+++ src/kimi_cli/app.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import contextlib
 import os
 import warnings
@@ -12,11 +14,25 @@ from kimi_cli.cli import InputFormat, OutputFormat
 from kimi_cli.config import LLMModel, LLMProvider, load_config
 from kimi_cli.llm import augment_provider_with_env_vars, create_llm
 from kimi_cli.session import Session
+from kimi_cli.share import get_share_dir
+from kimi_cli.soul import LLMNotSet, LLMNotSupported
 from kimi_cli.soul.agent import load_agent
 from kimi_cli.soul.context import Context
 from kimi_cli.soul.kimisoul import KimiSoul
 from kimi_cli.soul.runtime import Runtime
 from kimi_cli.utils.logging import StreamToLogger, logger
+
+
+def enable_logging(debug: bool = False) -> None:
+    if debug:
+        logger.enable("kosong")
+    logger.add(
+        get_share_dir() / "logs" / "kimi.log",
+        # FIXME: configure level for different modules
+        level="TRACE" if debug else "INFO",
+        rotation="06:00",
+        retention="10 days",
+    )
 
 
 class KimiCLI:
@@ -25,19 +41,18 @@ class KimiCLI:
         session: Session,
         *,
         yolo: bool = False,
-        stream: bool = True,  # TODO: remove this when we have a correct print mode impl
         mcp_configs: list[dict[str, Any]] | None = None,
         config_file: Path | None = None,
         model_name: str | None = None,
+        thinking: bool = False,
         agent_file: Path | None = None,
-    ) -> "KimiCLI":
+    ) -> KimiCLI:
         """
         Create a KimiCLI instance.
 
         Args:
             session (Session): A session created by `Session.create` or `Session.continue_`.
             yolo (bool, optional): Approve all actions without confirmation. Defaults to False.
-            stream (bool, optional): Use stream mode when calling LLM API. Defaults to True.
             config_file (Path | None, optional): Path to the configuration file. Defaults to None.
             model_name (str | None, optional): Name of the model to use. Defaults to None.
             agent_file (Path | None, optional): Path to the agent file. Defaults to None.
@@ -77,7 +92,7 @@ class KimiCLI:
         else:
             logger.info("Using LLM provider: {provider}", provider=provider)
             logger.info("Using LLM model: {model}", model=model)
-            llm = create_llm(provider, model, stream=stream, session_id=session.id)
+            llm = create_llm(provider, model, session_id=session.id)
 
         runtime = await Runtime.create(config, llm, session, yolo)
 
@@ -93,6 +108,10 @@ class KimiCLI:
             runtime,
             context=context,
         )
+        try:
+            soul.set_thinking(thinking)
+        except (LLMNotSet, LLMNotSupported) as e:
+            logger.warning("Failed to enable thinking mode: {error}", error=e)
         return KimiCLI(soul, runtime, env_overrides)
 
     def __init__(
@@ -127,7 +146,7 @@ class KimiCLI:
         finally:
             os.chdir(original_cwd)
 
-    async def run_shell_mode(self, command: str | None = None, markdown: bool = True) -> bool:
+    async def run_shell_mode(self, command: str | None = None) -> bool:
         from kimi_cli.ui.shell import ShellApp, WelcomeInfoItem
 
         welcome_info = [
@@ -139,6 +158,14 @@ class KimiCLI:
                 WelcomeInfoItem(
                     name="API URL",
                     value=f"{base_url} (from KIMI_BASE_URL)",
+                    level=WelcomeInfoItem.Level.WARN,
+                )
+            )
+        if self._env_overrides.get("KIMI_API_KEY"):
+            welcome_info.append(
+                WelcomeInfoItem(
+                    name="API Key",
+                    value="****** (from KIMI_API_KEY)",
                     level=WelcomeInfoItem.Level.WARN,
                 )
             )
@@ -154,7 +181,7 @@ class KimiCLI:
             welcome_info.append(
                 WelcomeInfoItem(
                     name="Model",
-                    value=f"{self._soul.model} (from KIMI_MODEL_NAME)",
+                    value=f"{self._soul.model_name} (from KIMI_MODEL_NAME)",
                     level=WelcomeInfoItem.Level.WARN,
                 )
             )
@@ -162,12 +189,12 @@ class KimiCLI:
             welcome_info.append(
                 WelcomeInfoItem(
                     name="Model",
-                    value=self._soul.model,
+                    value=self._soul.model_name,
                     level=WelcomeInfoItem.Level.INFO,
                 )
             )
         with self._app_env():
-            app = ShellApp(self._soul, welcome_info=welcome_info, markdown=markdown)
+            app = ShellApp(self._soul, welcome_info=welcome_info)
             return await app.run(command)
 
     async def run_print_mode(